style(schema-h2.sql) 스키마 id 타입 INT -> BIGINT 수정

+++ src/main/resources/sql/schema-h2.sql
@@ -29,19 +29,19 @@ CREATE TABLE IF NOT EXISTS theater
 
 CREATE TABLE IF NOT EXISTS auditorium
 (
-    id   INT AUTO_INCREMENT PRIMARY KEY,
-    theaterId  INT          NOT NULL,
-    name VARCHAR(255)       NOT NULL
+    id        BIGINT AUTO_INCREMENT PRIMARY KEY,
+    theaterId BIGINT       NOT NULL,
+    name      VARCHAR(255) NOT NULL
 );
 
 CREATE INDEX idx_auditorium_theater ON auditorium (theaterId);
 
 CREATE TABLE IF NOT EXISTS seat
 (
-    id         INT AUTO_INCREMENT PRIMARY KEY,
-    auditoriumId  INT          NOT NULL,
-    seatRow    VARCHAR(255) NOT NULL,
-    seatColumn INT          NOT NULL
+    id           BIGINT AUTO_INCREMENT PRIMARY KEY,
+    auditoriumId BIGINT       NOT NULL,
+    seatRow      VARCHAR(255) NOT NULL,
+    seatColumn   INT          NOT NULL
 );
 
 CREATE INDEX idx_seat_auditorium ON seat (auditoriumId);
@@ -49,13 +49,13 @@ CREATE INDEX idx_seat_row_column ON seat (seatRow, seatColumn);
 
 CREATE TABLE IF NOT EXISTS showing
 (
-    id        INT AUTO_INCREMENT PRIMARY KEY,
-    movieId   INT            NOT NULL,
-    auditoriumId INT            NOT NULL,
-    showDate  date           NOT NULL,
-    startTime time           NOT NULL,
-    endTime   time           NOT NULL,
-    price     decimal(10, 2) NOT NULL
+    id           BIGINT AUTO_INCREMENT PRIMARY KEY,
+    movieId      BIGINT         NOT NULL,
+    auditoriumId BIGINT         NOT NULL,
+    showDate     date           NOT NULL,
+    startTime    time           NOT NULL,
+    endTime      time           NOT NULL,
+    price        decimal(10, 2) NOT NULL
 );
 
 CREATE INDEX idx_showing_movie ON showing (movieId);
@@ -64,11 +64,11 @@ CREATE INDEX idx_showing_date ON showing (showDate);
 
 CREATE TABLE IF NOT EXISTS reservation
 (
-    id        INT AUTO_INCREMENT PRIMARY KEY,
-    memberId  INT       NOT NULL,
-    showingId INT       NOT NULL,
-    seatId    INT       NOT NULL,
-    createdAt  timestamp NOT NULL
+    id        BIGINT AUTO_INCREMENT PRIMARY KEY,
+    memberId  BIGINT    NOT NULL,
+    showingId BIGINT    NOT NULL,
+    seatId    BIGINT    NOT NULL,
+    createdAt timestamp NOT NULL
 );
 
 CREATE INDEX idx_reservation_member ON reservation (memberId);
